fix: refactor recover wallet
