+++ apps/mobile/src/features/recover-wallet/utils.ts
@@ -1,0 +1,8 @@
+import { isValidMnemonicWord } from '@leather.io/crypto';
+
+export function getInvalidMnemonicWords(recoveryMnemonic: string) {
+  return recoveryMnemonic
+    .trim()
+    .split(/\s+/g)
+    .filter(word => !isValidMnemonicWord(word));
+}
